+++ src/utils/mod.rs
@@ -1,0 +1,11 @@
+use crate::cli::Args;
+
+/// Prints messages only when verbose mode is enabled
+///
+/// This helper function is used throughout the application to conditionally
+/// print information messages based on the command-line arguments.
+pub fn verbose_print(args: &Args, message: impl AsRef<str>) {
+    if args.verbose && !args.silent {
+        println!("{}", message.as_ref());
+    }
+}
